update heroes ai dump info
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@2077 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/fheroes2/castle/captain.cpp
@@ -162,7 +162,7 @@ u8 Captain::GetControl(void) const
     return home.GetControl();
 }
 
-u16 Captain::GetIndex(void) const
+s32 Captain::GetIndex(void) const
 {
     return home.GetIndex();
 }
